0-print_list.c

+++ 0x12-singly_linked_lists/0-print_list.c
@@ -1,5 +1,4 @@
 #include <stdio.h>
-#include <stdlib.h>
 #include "lists.h"
 
 /**
@@ -11,7 +10,7 @@
 size_t print_list(const list_t *h)
 {
 	const list_t *ptr = h;
-	int size = 0;
+	size_t size = 0;
 
 	if (ptr == NULL)
 		return (0);
